modestas_vatican_records/pyproject.toml
[tool.poetry]
name = "project-rag"
version = "0.1.0"
description = "Basic RAG application"
authors = ["Modestas Grazys"]
readme = "README.md"
packages = [{include = "chatbot"}]

[tool.poetry.dependencies]
python = ">=3.12,<4.0"
pydantic = ">=2.10.6,<3.0.0"
chainlit = ">=2.2.1,<3.0.0"
chromadb = ">=0.6.3,<0.7.0"
python-dotenv = "==1.0.1"
uuid = ">=1.30,<2.0"
pathlib = ">=1.0.1,<2.0.0"
langchain = ">=0.3.19,<0.4.0"
asyncio = ">=3.4.3,<4.0.0"
transformers = ">=4.49.0,<5.0.0"
numpy = "==1.26.4"
sentence-transformers = ">=3.4.1,<4.0.0"
torch = "2.6.0"
torchvision = "0.21.0"
google = "^3.0.0"
google-generativeai = "^0.8.4"
tqdm = "^4.67.1"
huggingface-hub = "^0.29.1"
accelerate = "^1.4.0"
matplotlib = "^3.10.1"
umap-learn = "^0.5.7"
pypdf = "^5.4.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
